refactored Proxy class
