+++ OnlinerTracker.Api/App_Start/AutofacConfig.cs
@@ -22,12 +22,7 @@ namespace OnlinerTracker.Api
             builder.RegisterType<AuthorizationService>().As<IAuthorizationService>().InstancePerDependency();
             builder.RegisterType<ProductService>().As<IProductService>().InstancePerDependency();
             builder.Register(c => new TrackingService()).As<ITrackingService>().InstancePerDependency();
-            builder.Register(c => new ExternalProductProxy(c.Resolve<IProductService>())
-                                    {
-                                        RemoteServiceUrl = ConfigurationManager.AppSettings["remoteService:Url"],
-                                        RemoteServiceType = ConfigurationManager.AppSettings["remoteService:Type"]
-                                    })
-                    .As<IExternalProductService>().InstancePerDependency();
+            builder.Register(c => new OnlinerProxy(c.Resolve<IProductService>())).As<IExternalProductService>().InstancePerDependency();
 
             #endregion
