+++ OnlinerTracker.Api/Web.config
@@ -13,10 +13,6 @@
     <add name="DataConnection" providerName="System.Data.SqlClient" connectionString="Server=KARTYNNIKA;Database=OnlinerTracker;Integrated Security=True" />
   </connectionStrings>
   <appSettings>
-    <add key="remoteService:Url" value="https://catalog.api.onliner.by/search/products" />
-    <!-- Remote services types we use: 0 = www.oliner.by -->
-    <add key="remoteService:Type" value="0" /> 
-    
     <add key="authConfig:GoogleClientId" value="662974838096-ea4had89gn6gr6edt7pkfm2drrjrjhcj.apps.googleusercontent.com" />
     <add key="authConfig:GoogleClientSecret" value="HuzdYDwm-RTXFoFSRyKqQxb5" />
     <add key="authConfig:TwitterConsumerKey" value="am9fD3Y9QxyP3dcfz0ISQqNdY" />
