+++ OnlinerTracker.Data/OnlinerTracker.Data.csproj
@@ -63,7 +63,6 @@
     <Compile Include="Product.cs" />
     <Compile Include="ExternalProduct.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RemoteServiceType.cs" />
     <Compile Include="SecurityModels\Client.cs" />
     <Compile Include="SecurityModels\ApplicationType.cs" />
     <Compile Include="SecurityModels\ExternalLoginData.cs" />
