--- OnlinerTracker.Data/RemoteServiceType.cs
@@ -1,7 +1,0 @@
-﻿namespace OnlinerTracker.Data
-{
-    public enum RemoteServiceType
-    {
-        Onliner = 0
-    }
-}
