+++ OnlinerTracker.Proxies/OnlinerTracker.Proxies.csproj
@@ -53,7 +53,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ExternalProductProxy.cs" />
+    <Compile Include="OnlinerProxy.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
